requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
Django==4.0.3
django-filter==2.4.0
djangorestframework==3.13.1
Pillow==9.1.0
psycopg2==2.9.2
pytz==2022.1
simplejson==3.17.6
sqlparse==0.4.2
django-cors-headers==3.11.0

